chore(TU-15133): Update NextJS demos to run on latest version
Add examples for the App Router available from NextJS version >= 13.
