+++ package.json
@@ -57,7 +57,7 @@
     "jest-transform-css": "^2.1.0",
     "lerna": "^3.22.1",
     "mini-css-extract-plugin": "^1.3.7",
-    "next": "12.1.0",
+    "next": "^14.2.5",
     "prettier": "^2.2.1",
     "prop-types": "^15.7.2",
     "raw-loader": "^4.0.2",
