--- packages/demo-nextjs/components/flex.js
@@ -1,7 +1,0 @@
-const Flex = () => (
-  <span role="img" aria-label="flex">
-    💪
-  </span>
-)
-
-export default Flex
