--- packages/demo-nextjs/components/sparkle.js
@@ -1,7 +1,0 @@
-const Sparkle = () => (
-  <span role="img" aria-label="sparkle">
-    ✨
-  </span>
-)
-
-export default Sparkle
